+++ packages/zowe-explorer/src/shared/refresh.ts
@@ -12,12 +12,13 @@
 import { IZoweTree, IZoweTreeNode } from "@zowe/zowe-explorer-api";
 import { PersistentFilters } from "../PersistentFilters";
 import { Profiles } from "../Profiles";
-import { removeSession, syncSessionNode } from "../utils/ProfilesUtils";
+import { syncSessionNode } from "../utils/ProfilesUtils";
 import { ZoweExplorerApiRegister } from "../ZoweExplorerApiRegister";
 import { resetValidationSettings, returnIconState } from "./actions";
 import { labelRefresh } from "./utils";
 import * as contextually from "../shared/context";
 import * as globals from "../globals";
+import { removeSession } from "../utils/SessionUtils";
 
 /**
  * View (DATA SETS, JOBS, USS) refresh button
